split lvl_script_value off from main file
